+++ requirements.txt
@@ -2,5 +2,6 @@ awscli
 botocore
 boto
 boto3
+configparser
 invoke
 requests
