+++ tasks/env.py
@@ -9,6 +9,8 @@ FAASM_HOME = join(HOME_DIR, "faasm")
 
 ANSIBLE_DIR = join(PROJ_ROOT, "ansible")
 
+WASM_FUNC_BUILD_DIR = join(PROJ_ROOT, "func", "wasm_func_build")
+
 WASM_DIR = join(PROJ_ROOT, "wasm")
 
 RUNTIME_S3_BUCKET = "faasm-runtime"
